requirements.txt
streamlit
google-genai
groq
requests
